+++ src/podofo/main/PdfPage.h
@@ -21,6 +21,7 @@ namespace PoDoFo {
 
 class PdfDocument;
 class InputStream;
+class PdfPage;
 
 struct PODOFO_API PdfTextEntry final
 {
@@ -37,6 +38,86 @@ struct PODOFO_API PdfTextExtractParams final
     nullable<Rect> ClipRect;
     PdfTextExtractFlags Flags = PdfTextExtractFlags::None;
 };
+
+template <typename TField>
+class PdfPageFieldIterableBase final
+{
+    friend class PdfPage;
+
+public:
+    PdfPageFieldIterableBase()
+        : m_page(nullptr) { }
+
+private:
+    PdfPageFieldIterableBase(PdfPage& page)
+        : m_page(&page) { }
+
+public:
+    class Iterator final
+    {
+        friend class PdfPageFieldIterableBase;
+    public:
+        using difference_type = void;
+        using value_type = TField*;
+        using pointer = void;
+        using reference = void;
+        using iterator_category = std::forward_iterator_tag;
+    public:
+        Iterator()
+            : m_Field(nullptr) { }
+    private:
+        void stepIntoPageAnnot();
+
+        Iterator(PdfAnnotationCollection::iterator begin,
+            PdfAnnotationCollection::iterator end)
+            : m_annotsIterator(std::move(begin)), m_annotsEnd(std::move(end)), m_Field(nullptr)
+        {
+            stepIntoPageAnnot();
+        }
+
+    public:
+        Iterator(const Iterator&) = default;
+        Iterator& operator=(const Iterator&) = default;
+        bool operator==(const Iterator& rhs) const
+        {
+            return m_annotsIterator == rhs.m_annotsIterator;
+        }
+        bool operator!=(const Iterator& rhs) const
+        {
+            return m_annotsIterator != rhs.m_annotsIterator;
+        }
+        Iterator& operator++()
+        {
+            m_annotsIterator++;
+            stepIntoPageAnnot();
+            return *this;
+        }
+        Iterator operator++(int)
+        {
+            auto copy = *this;
+            m_annotsIterator++;
+            stepIntoPageAnnot();
+            return copy;
+        }
+        value_type operator*() { return m_Field; }
+        value_type operator->() { return m_Field; }
+    private:
+        PdfAnnotationCollection::iterator m_annotsIterator;
+        PdfAnnotationCollection::iterator m_annotsEnd;
+        value_type m_Field;
+        std::unordered_set<PdfReference> m_visitedObjs;
+    };
+
+public:
+    Iterator begin() const;
+    Iterator end() const;
+
+private:
+    PdfPage* m_page;
+};
+
+using PdfPageFieldIterable = PdfPageFieldIterableBase<PdfField>;
+using PdfPageConstFieldIterable = PdfPageFieldIterableBase<const PdfField>;
 
 /** PdfPage is one page in the pdf document.
  *  It is possible to draw on a page using a PdfPainter object.
@@ -187,6 +268,13 @@ public:
     void SetICCProfile(const std::string_view& csTag, InputStream& stream, int64_t colorComponents,
         PdfColorSpaceType alternateColorSpace = PdfColorSpaceType::DeviceRGB);
 
+    /**
+     * Get an iterator for all fields in the page. All widget annotation fields
+     * in the pages will be returned
+     */
+    PdfPageFieldIterable GetFieldsIterator();
+    PdfPageConstFieldIterable GetFieldsIterator() const;
+
 public:
     unsigned GetIndex() const { return m_Index; }
     PdfContents& GetOrCreateContents();
@@ -265,6 +353,50 @@ template<typename TField>
 TField& PdfPage::CreateField(const std::string_view& name, const Rect & rect, bool rawRect)
 {
     return static_cast<TField&>(createField(name, typeid(TField), rect, rawRect));
+}
+
+template<typename TField>
+typename PdfPageFieldIterableBase<TField>::Iterator PdfPageFieldIterableBase<TField>::begin() const
+{
+    if (m_page == nullptr)
+        return Iterator();
+    else
+        return Iterator(m_page->GetAnnotations().begin(), m_page->GetAnnotations().end());
+}
+
+template<typename TField>
+typename PdfPageFieldIterableBase<TField>::Iterator PdfPageFieldIterableBase<TField>::end() const
+{
+    if (m_page == nullptr)
+        return Iterator();
+    else
+        return Iterator(m_page->GetAnnotations().end(), m_page->GetAnnotations().end());
+}
+
+template<typename TField>
+void PdfPageFieldIterableBase<TField>::Iterator::stepIntoPageAnnot()
+{
+    while (true)
+    {
+        if (m_annotsIterator == m_annotsEnd)
+            break;
+
+        auto& annot = **m_annotsIterator;
+        PdfField* field = nullptr;
+        if (annot.GetType() == PdfAnnotationType::Widget &&
+            (field = &static_cast<PdfAnnotationWidget&>(annot).GetField(),
+                m_visitedObjs.find(field->GetObject().GetIndirectReference()) == m_visitedObjs.end()))
+        {
+            m_Field = field;
+            m_visitedObjs.insert(field->GetObject().GetIndirectReference());
+            return;
+        }
+
+        m_annotsIterator++;
+    }
+
+    m_Field = nullptr;
+    m_visitedObjs.clear();
 }
 
 };