+++ CHANGELOG.md
@@ -6,6 +6,7 @@
 PdfOutlineItem, PdfOutlines, PdfAnnotationActionBase, PdfAnnotationLink PdfAnnotationFileAttachment
 - Reviewed PdfNameTree
 - PdfDocument: Added GetFieldsIterator()
+- PdfPage: Added GetFieldsIterator()
 
 ## Version 0.10.3
 - Fixed big performance regression introduced in 0.10, see #108